+++ src/app/context.rs
@@ -1,10 +1,13 @@
-use crate::{
-    cli::AppLaunchArgs,
-    conf::Conf,
-    display::{Cols, DEFAULT_COLS},
-    pattern::SearchModeMap,
-    tree::SpecialPath,
-    verb::VerbStore,
+use {
+    super::*,
+    crate::{
+        cli::AppLaunchArgs,
+        conf::Conf,
+        display::{Cols, DEFAULT_COLS},
+        pattern::SearchModeMap,
+        tree::SpecialPath,
+        verb::VerbStore,
+    },
 };
 
 /// The immutable container that can be passed around
@@ -28,6 +31,8 @@ pub struct AppContext {
 
     /// order of columns in tree display
     pub cols: Cols,
+
+    pub standard_status: StandardStatus,
 }
 
 impl AppContext {
@@ -37,6 +42,7 @@ impl AppContext {
         config: &Conf,
     ) -> Self {
         let config_path = Conf::default_location().to_string_lossy().to_string();
+        let standard_status = StandardStatus::new(&verb_store);
         Self {
             config_path,
             launch_args,
@@ -44,6 +50,7 @@ impl AppContext {
             special_paths: config.special_paths.clone(),
             search_modes: config.search_modes.clone(),
             cols: config.cols_order.unwrap_or(DEFAULT_COLS).clone(),
+            standard_status,
         }
     }
 }